% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInitialModel.R
\name{createInitialModel}
\alias{createInitialModel}
\title{Create Initial Model Configuration}
\usage{
createInitialModel(n_factors, n_items, name_items, exclude_items = NULL)
}
\arguments{
\item{n_factors}{Number of factors.}

\item{n_items}{Number of items.}

\item{name_items}{Vector of item names or prefix.}

\item{exclude_items}{Optional vector of items to exclude.}
}
\value{
A list with model configuration parameters.
}
\description{
Creates an initial model configuration list for factor analysis.
}
