% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAllResults.R
\name{extractAllResults}
\alias{extractAllResults}
\title{Extract All Results from Bootstrap Analysis}
\usage{
extractAllResults(res)
}
\arguments{
\item{res}{A bootstrap result object containing processed_results.}
}
\value{
A data frame with results from all replications and iterations.
}
\description{
Extracts and combines all results from bootstrap replications.
}
