% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_metricas.R
\name{extraer_metricas}
\alias{extraer_metricas}
\title{Extract Fit Metrics from Model}
\usage{
extraer_metricas(modelo_fit)
}
\arguments{
\item{modelo_fit}{A model fit object with Measure and Value columns.}
}
\value{
A data frame with formatted fit metrics.
}
\description{
Extracts and formats common fit metrics (AIC, BIC, CFI, TLI, RMSEA) from a model.
}
