% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_decision_SSV.R
\name{interpret_decision_SSV}
\alias{interpret_decision_SSV}
\title{Interpret Modification Index Decisions}
\usage{
interpret_decision_SSV(MI_Saris)
}
\arguments{
\item{MI_Saris}{Data frame with modification indices and decision.pow column.}
}
\value{
A data frame with interpretations of each decision.
}
\description{
Interprets modification indices and power decisions from SEM analysis.
}
