% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_per_fit_efa.R
\name{calculate_per_fit_efa}
\alias{calculate_per_fit_efa}
\title{Calculate Percentage of Bootstrap EFA Models Meeting Fit Thresholds}
\usage{
calculate_per_fit_efa(boot_efa_results, thresholds_str)
}
\arguments{
\item{boot_efa_results}{Results from boot_efa function.}

\item{thresholds_str}{A string specifying thresholds in format
"MEASURE DIRECTION VALUE" separated by commas.
Example: "CFI > 0.95, TLI > 0.95, RMSEA < 0.06, SRMR < 0.08"}
}
\value{
A data frame with percentages for each threshold.
}
\description{
Calculates the percentage of bootstrap EFA replications that meet
specified fit measure thresholds.
}
\examples{
\donttest{
set.seed(123)
n <- 200
data <- as.data.frame(lapply(setNames(1:6, paste0("PBA", 1:6)),
                             function(i) sample(1:5, n, replace = TRUE)))
results_boot_efa <- boot_efa(
  data = data, n_factors = 2, n_items = 6, name_items = "PBA",
  rotation = "oblimin", n_replications = 5
)
thresholds <- "CFI > 0.95, TLI > 0.95, RMSEA < 0.06, SRMR < 0.08"
calculate_per_fit_efa(results_boot_efa, thresholds)
}
}
