% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_afe.R
\name{cor_afe}
\alias{cor_afe}
\title{Extract Factor Correlations from EFA}
\usage{
cor_afe(x)
}
\arguments{
\item{x}{An EFA result object containing factor correlations.}
}
\value{
A matrix of factor correlations rounded to 2 decimals.
}
\description{
Extracts the inter-factor correlation matrix (Phi) from EFA results.
}
