\name{EFA_modern}
\alias{EFA_modern}
\title{Perform Exploratory Factor Analysis (EFA) with Modern Specifications}

\description{
The `EFA_modern` function facilitates the execution of Exploratory Factor Analysis (EFA) using the lavaan package. It generates a model based on specified factors and items, computes model fit indices, standardizes solutions, and assesses inter-factor correlations, returning all relevant outputs in a structured list.
}

\usage{
EFA_modern(n_factors, n_items, name_items, data, apply_threshold, estimator = "WLSMV", exclude_items = NULL)
}

\arguments{
  \item{n_factors}{An integer indicating the number of factors to be extracted in the EFA model.}
  \item{n_items}{An integer specifying the number of items included in the analysis.}
  \item{name_items}{A string representing the base name of the items included in the model (e.g., "IS" for IS_1, IS_2, etc.).}
  \item{data}{A data frame containing the data to be used for the EFA.}
  \item{apply_threshold}{A logical value indicating whether to apply thresholds in the standardized solutions.}
  \item{estimator}{A string specifying the estimator to be used in the EFA. Default is "WLSMV".}
  \item{exclude_items}{A vector of item names or indices to be excluded from the analysis. Default is \code{NULL}, meaning no items are excluded.}
}

\details{
The `EFA_modern` function is designed to simplify the process of conducting Exploratory Factor Analysis using a modern approach with the lavaan package. The function handles the generation of EFA models, applies specified exclusions, computes key fit indices, and standardizes the solution matrix. Additionally, it calculates inter-factor correlations, providing a comprehensive output that is ready for further interpretation or reporting.

The function is highly flexible, allowing the user to specify the number of factors, items, and items to be excluded. The default estimator is "WLSMV", which is appropriate for models with categorical data, but this can be modified to suit different data types.

Messages and warnings are managed within the function to ensure clean outputs. The resulting list includes the fit indices, specifications, inter-factor correlations, and the standardized solution matrix, all of which are essential for interpreting the results of the EFA.
}

\value{
A \code{list} containing the following elements:
\itemize{
  \item \code{Bondades_Original} - A list of model fit indices extracted from the EFA.
  \item \code{Specifications} - The specifications of the EFA model as generated by the lavaan package.
  \item \code{InterFactor} - A matrix of inter-factor correlations derived from the EFA model.
  \item \code{result_df} - The standardized solution matrix of the EFA, with optional thresholds applied.
}
}

\examples{
# Example usage of the EFA_modern function

# Perform EFA with 4 factors and 21 items
results <- EFA_modern(n_factors = 4, n_items = 21, name_items = "IS", data = df_EFA, apply_threshold = TRUE)

# Access the model fit indices
Bondades_Original <- results$Bondades_Original

# View the standardized solution matrix
result_df <- results$result_df
print(result_df, n = 21)

# Examine inter-factor correlations
InterFactor <- results$InterFactor
print(InterFactor)
}

\author{
Dr. José Ventura-León
}
