% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFA_plot.R
\name{EFA_plot}
\alias{EFA_plot}
\title{Plot EFA Factor Loadings}
\usage{
EFA_plot(specifications, item_prefix)
}
\arguments{
\item{specifications}{Lavaan model object or a data.frame with columns:
est.std, ci.lower, ci.upper, lhs, rhs, op.}

\item{item_prefix}{Prefix for item names.}
}
\value{
A ggplot object showing factor loadings.
}
\description{
Creates a visualization of factor loadings from EFA.
}
