% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bondad_boot_AFC.R
\name{extract_bondad_boot_AFC}
\alias{extract_bondad_boot_AFC}
\title{Extract Fit Measures from Bootstrap CFA Results}
\usage{
extract_bondad_boot_AFC(results)
}
\arguments{
\item{results}{A list of CFA results containing FitMeasuresDf.}
}
\value{
A combined data frame of fit measures with sample and iteration identifiers.
}
\description{
Extracts and combines fit measures from bootstrap CFA analyses.
}
