% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bondad_boot_EFA.R
\name{extract_bondad_boot_EFA}
\alias{extract_bondad_boot_EFA}
\title{Extract Fit Measures from Bootstrap EFA Results}
\usage{
extract_bondad_boot_EFA(resultados_bootstrap)
}
\arguments{
\item{resultados_bootstrap}{A bootstrap result object with Results list.}
}
\value{
A data frame of fit measures with sample identifiers.
}
\description{
Extracts and combines fit measures from bootstrap EFA analyses.
}
