\name{extract_fit}
\alias{extract_fit}
\title{Extraction of Goodness-of-Fit Measures from Psych Objects}

\description{
This function is specifically designed to extract and present key goodness-of-fit indices from an object created by factor analysis functions in the `psych` package. It compiles important fit measures such as Root Mean Square Error of Approximation (RMSEA), Tucker-Lewis Index (TLI), Comparative Fit Index (CFI), and Bayesian Information Criterion (BIC), providing a clear summary of how well the model fits the data.
}

\usage{
extract_fit(factors_data)
}

\arguments{
  \item{factors_data}{An object generated by a factor analysis function from the `psych` package. This object should contain various goodness-of-fit measures resulting from the factor analysis.}
}

\details{
The function calculates the Comparative Fit Index (CFI) internally and extracts other fit indices directly from the `factors_data` object. These indices include RMSEA and its confidence interval (lower and upper bounds), TLI, CFI, null chi-squared value, the objective function value, and BIC. The CFI is computed based on the provided statistics in the `factors_data` object.
}

\value{
A data frame with two columns: `Bondades`, listing the names of the extracted fit indices, and `values`, containing the corresponding values for these indices.
}

\examples{
# Assuming 'factors_data' is an object resulting from a factor analysis in the psych package
fit_summary <- extract_fit(factors_data)
print(fit_summary)
}

\author{
Professor José Ventura León
}
