% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit_measures.R
\name{extract_fit_measures}
\alias{extract_fit_measures}
\title{Extract Fit Measures from Lavaan Specifications}
\usage{
extract_fit_measures(Specifications)
}
\arguments{
\item{Specifications}{A list of lavaan model objects.}
}
\value{
A data frame with fit measures for each factor solution.
}
\description{
Extracts fit measures (chi-square, SRMR, WRMR, CFI, TLI, RMSEA) from lavaan models.
}
