% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ModIndex_AFC.R
\name{extract_ModIndex_AFC}
\alias{extract_ModIndex_AFC}
\title{Extract Modification Indices from CFA Results}
\usage{
extract_ModIndex_AFC(results)
}
\arguments{
\item{results}{A list of CFA results containing ModificationsDf.}
}
\value{
A combined data frame of modification indices with sample identifiers.
}
\description{
Extracts and combines modification indices from multiple CFA analyses.
}
