% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_lambda.R
\name{Matrix_lambda}
\alias{Matrix_lambda}
\title{Create Lambda Matrix for CFA}
\usage{
Matrix_lambda(n_items, factores, n_tamanos)
}
\arguments{
\item{n_items}{Total number of items.}

\item{factores}{Number of factors.}

\item{n_tamanos}{Vector specifying the number of items per factor.}
}
\value{
A matrix with 1s indicating item-factor loadings.
}
\description{
Creates a binary lambda matrix specifying item-factor relationships for CFA models.
}
