\name{plot_cfa_stability}
\alias{plot_cfa_stability}
\title{Plot Stability of CFA Model Fit Indices Across Varying Sample Sizes}

\description{
The `plot_cfa_stability` function generates a series of line plots to visualize the stability of fit indices and reliability measures for a Confirmatory Factor Analysis (CFA) model across different sample sizes. The function allows for customization of axis limits, breaks, and the appearance of the plots.
}

\usage{
plot_cfa_stability(resultados,
                   y_min_cfi = 0.9, y_max_cfi = 1,
                   y_min_rmsea = 0, y_max_rmsea = 0.15,
                   y_min_srmr = 0, y_max_srmr = 0.15,
                   y_min_crmr = 0, y_max_crmr = 0.15,
                   y_min_reliability = 0.50, y_max_reliability = 1,
                   y_breaks = 0.01,
                   y_breaks_reliability = 0.05,
                   hline_color = "red",
                   xlab_size = 12,
                   ylab_size = 12,
                   label_size = 10)
}

\arguments{
  \item{resultados}{A dataframe containing the CFA model fit indices and reliability measures, typically generated by the `boot_cfa_stability` function.}
  \item{y_min_cfi}{A numeric value indicating the minimum y-axis limit for the CFI plot. Default is 0.9.}
  \item{y_max_cfi}{A numeric value indicating the maximum y-axis limit for the CFI plot. Default is 1.}
  \item{y_min_rmsea}{A numeric value indicating the minimum y-axis limit for the RMSEA plot. Default is 0.}
  \item{y_max_rmsea}{A numeric value indicating the maximum y-axis limit for the RMSEA plot. Default is 0.15.}
  \item{y_min_srmr}{A numeric value indicating the minimum y-axis limit for the SRMR plot. Default is 0.}
  \item{y_max_srmr}{A numeric value indicating the maximum y-axis limit for the SRMR plot. Default is 0.15.}
  \item{y_min_crmr}{A numeric value indicating the minimum y-axis limit for the CRMR plot. Default is 0.}
  \item{y_max_crmr}{A numeric value indicating the maximum y-axis limit for the CRMR plot. Default is 0.15.}
  \item{y_min_reliability}{A numeric value indicating the minimum y-axis limit for the reliability plot. Default is 0.50.}
  \item{y_max_reliability}{A numeric value indicating the maximum y-axis limit for the reliability plot. Default is 1.}
  \item{y_breaks}{A numeric value specifying the interval between y-axis breaks for the CFI, RMSEA, SRMR, and CRMR plots. Default is 0.01.}
  \item{y_breaks_reliability}{A numeric value specifying the interval between y-axis breaks for the reliability plot. Default is 0.05.}
  \item{hline_color}{A string specifying the color of the horizontal reference lines in the plots. Default is "red".}
  \item{xlab_size}{A numeric value determining the size of the x-axis labels. Default is 12.}
  \item{ylab_size}{A numeric value determining the size of the y-axis labels. Default is 12.}
  \item{label_size}{A numeric value determining the size of the labels for the subplots. Default is 10.}
}

\details{
The `plot_cfa_stability` function creates multiple plots to assess the stability of various CFA model fit indices (CFI, RMSEA, SRMR, CRMR) and a reliability measure across different percentages of the original sample size.

Each plot displays the mean value of the fit index or reliability measure as a function of the percentage of the original sample size. Horizontal reference lines are included to indicate typical cutoff values for model fit and reliability.

The function returns a combined ggplot object, arranging all the individual plots into a single figure. The figure provides a visual summary of how the model's performance varies with changes in sample size.
}

\value{
A \code{ggplot} object representing the combined plots of CFA model fit indices and reliability across varying sample sizes. The object includes:
\itemize{
  \item A line plot for the CFI index across sample sizes.
  \item A line plot for the RMSEA index across sample sizes.
  \item A line plot for the SRMR index across sample sizes.
  \item A line plot for the CRMR index across sample sizes.
  \item A line plot for the reliability measure across sample sizes.
}
}

\examples{
# Example usage of the plot_cfa_stability function

# Load necessary libraries
library(ggplot2)
library(tidyverse)

resultado <- boot_cfa_stability(model_original,
                              data = df,
                              seed = 2023,
                              num_replicas = 1000,
                              estimator = "WLSMV",
                              n_cores = 10)

resultado

# Generate the stability plots
figure4 <- plot_cfa_stability(resultado,
                              y_min_cfi = 0.9, y_max_cfi = 1,
                              y_min_rmsea = 0, y_max_rmsea = 0.15,
                              y_min_srmr = 0, y_max_srmr = 0.15,
                              y_min_crmr = 0, y_max_crmr = 0.15,
                              y_min_reliability = 0.50, y_max_reliability = 1,
                              y_breaks = 0.01,
                              y_breaks_reliability = 0.05,
                              hline_color = "red",
                              xlab_size = 12,
                              ylab_size = 12,
                              label_size = 10)

# Display the plots
print(figure4)
}

\author{
Dr. José Ventura-León
}
