% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preliminar_table.R
\name{Preliminar_table}
\alias{Preliminar_table}
\title{Preliminary Response Frequency Table}
\usage{
Preliminar_table(data)
}
\arguments{
\item{data}{Data frame containing the items.}
}
\value{
A tibble with response frequencies as percentages.
}
\description{
Creates a table of response frequencies for items.
}
