% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsyMetricTools-package.R
\docType{package}
\name{PsyMetricTools-package}
\alias{PsyMetricTools}
\alias{PsyMetricTools-package}
\title{PsyMetricTools: Psychometric and Statistical Analysis Tools}
\description{
Provides tools for psychometric and statistical analysis in social sciences, including functions for data preprocessing, factor analysis, reliability testing, descriptive statistics, visualization of Likert-scale items, measurement invariance analysis, and handling of multi-class imbalance. Facilitates advanced data analysis tasks commonly encountered in psychological and educational research. Includes an updated easy_invariance() function for ordinal data following Wu and Estabrook (2016) \doi{10.1037/met0000051}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jventural/PsyMetricTools}
  \item Report bugs at \url{https://github.com/jventural/PsyMetricTools/issues}
}

}
\author{
\strong{Maintainer}: Jose Ventura-Leon \email{jventuraleon@gmail.com} (\href{https://orcid.org/0000-0003-2996-4244}{ORCID})

}
\keyword{internal}
