\name{specification_models}
\alias{specification_models}
\title{Factorial Model Specification Tool Using Lavaan}

\description{
This function is designed to facilitate the specification and estimation of exploratory factorial models using the lavaan package. It takes pre-defined model formulas and applies them to a dataset to estimate the model parameters. This function is particularly useful for conducting confirmatory factor analysis (CFA) in psychological and social science research, allowing for the evaluation of various factorial structures.
}

\usage{
specification_models(modelos, data, estimator, rotation = "oblimin", verbose = FALSE)
}

\arguments{
  \item{modelos}{A list of model specifications generated by the generate_models function. Each model specification should be a string representing the lavaan model syntax.}
  \item{data}{The dataset to which the model will be applied. This should be a data frame with variables used in the model specifications.}
  \item{estimator}{The estimation method to be used in the confirmatory factor analysis. Common choices include ML (Maximum Likelihood), WLSMV (Weighted Least Squares Mean and Variance adjusted), among others.}
  \item{rotation}{The rotation method to use for EFA (default "oblimin").}
  \item{verbose}{Logical, whether to print detailed output (default FALSE).}
}

\details{
The function iterates over each model in the modelos list, applying each model specification to the given data using the specified estimator. It estimates each model using lavaan and returns the fitted model objects. This allows for the subsequent extraction of fit indices, factor loadings, and other relevant statistics.
}

\value{
A list of fitted lavaan model objects. Each object contains the results of the confirmatory factor analysis for one of the specified models, including estimates of parameters, fit indices, and diagnostics.
}

\examples{
# Assuming modelos is a list of model specifications and Data is the dataset
Specifications <- specification_models(modelos, data = Data, estimator = "WLSMV")
print(Specifications)
}

\author{
Dr. Jose Ventura-Leon
}

